+++ src/js/view/room/modal/settings.mjs
@@ -29,6 +29,7 @@ import {
   deleteRoom,
   getTableNo,
   setTableNo,
+  getServerPreference,
   getRoomPreference,
   setRoomPreference
 } from '../../../state/index.mjs'
@@ -243,7 +244,7 @@ export function modalSettings () {
   for (let i = 0; i < room.backgrounds.length; i++) {
     const option = _('option').create(room.backgrounds[i].name)
     option.value = i
-    if (i === getRoomPreference('background', 0)) option.selected = true
+    if (i === getServerPreference('background', 0)) option.selected = true
     backgrounds.add(option)
   }