+++ CHANGELOG.md
@@ -11,10 +11,11 @@ Download current and previous versions from [https://github.com/ludus-leonis/Fre
 ### Other changes
 
 * added separate color sets in templates for piece borders and backgrounds
-* fixed "RPG" template glass tiles transparency
-* fixed piece.h default handling (now equals w, not 0)
 * added unit test docs and added `Dockerfile` for test webservers
 * changed all JS-CSS injections to use custom properties
+* changed background settings from room to server setting
+* fixed "RPG" template glass tiles transparency
+* fixed piece.h default handling (now equals w, not 0)
 
 ## v0.12.0 - Hexing Heron